Don't leave haproxy behind upon stop
This PR changes few things around the way a deployment terminates. First
of all it moves some PID files into /var/run/appscale: this allows for
easily terminating stray process from terminate.rb. Second it adds logic
to start gentle termination from the controller itself: since the AC
starts the process, it should have a better understanding of the order
of service termination.

+++ AppController/lib/haproxy.rb
@@ -31,7 +31,7 @@ module HAProxy
   SERVICE_SITES_PATH = File.join(HAPROXY_PATH, 'service-sites-enabled')
   SERVICE_MAIN_FILE = File.join(HAPROXY_PATH, "service-haproxy.#{CONFIG_EXTENSION}")
   SERVICE_BASE_FILE = File.join(HAPROXY_PATH, "service-base.#{CONFIG_EXTENSION}")
-  SERVICE_PIDFILE = '/var/run/service-haproxy.pid'.freeze
+  SERVICE_PIDFILE = '/var/run/appscale/service-haproxy.pid'.freeze
 
   # Maximum AppServer threaded connections
   MAX_APPSERVER_CONN = 7
@@ -78,11 +78,21 @@ module HAProxy
 
   # Start HAProxy for API services.
   def self.services_start
+    if !valid_config?(SERVICE_MAIN_FILE)
+      Djinn.log_warn('Invalid configuration for HAProxy services.')
+      return
+    end
+    return if MonitInterface.is_running?(:service_haproxy)
+
     start_cmd = "#{HAPROXY_BIN} -f #{SERVICE_MAIN_FILE} -D " \
       "-p #{SERVICE_PIDFILE}"
     stop_cmd = "#{BASH_BIN} -c 'kill $(cat #{SERVICE_PIDFILE})'"
     MonitInterface.start_daemon(
       :service_haproxy, start_cmd, stop_cmd, SERVICE_PIDFILE)
+  end
+
+  def self.services_stop
+    MonitInterface.stop_daemon(:service_haproxy)
   end
 
   # Create the config file for UserAppServer.
@@ -132,7 +142,7 @@ module HAProxy
     # We only serve internal services here.
     unless [TaskQueue::NAME, DatastoreServer::NAME,
         UserAppClient::NAME, BlobServer::NAME].include?(name)
-      Djinn.log_warn('create_app_config called for unknown service: #{name}.')
+      Djinn.log_warn("create_app_config called for unknown service: #{name}.")
       return false
     end
